fied

+++ app_heating_calc/models.py
@@ -11,8 +11,8 @@ class  Region(models.Model):
 class District(models.Model):
     region =  models.ForeignKey(Region, on_delete=models.CASCADE)
     name  =  models.CharField(max_length=200)
-    average_temperature = models.IntegerField()
-    time =  models.IntegerField()
+    average_temperature = models.FloatField()
+    time =  models.FloatField()
 
     def __str__(self):
         return self.name
